Added proper JavaDoc documentation (at least I think I did), and did a little fixing with the Graph.java, HashGraph.java, and TreeGraph.java

+++ samminhch/Maths.java
@@ -2,11 +2,17 @@ package samminhch;
 
 import samminhch.utils.StringUtils;
 
+/**
+ * The Maths class contains functions for Math functions that I wish Java had.
+ *
+ * @since September 16, 2021
+ */
 public class Maths {
     /**
      * Returns a decimal integer number in scientific notation. This is literally useless you could just use
      * String.format() for this.
-     * @param num the number you want to convert to Scientific Notation
+     *
+     * @param num            the number you want to convert to Scientific Notation
      * @param decimalsToKeep the number of decimals you want to keep in the conversion
      * @return the number in scientific notation
      * @deprecated Just use String.format ;-;
@@ -20,7 +26,14 @@ public class Maths {
                 "e+" + (Integer.toString(num).length() - 1);
     }
 
-    //Please fix this method this is so shit lol
+    /**
+     * This function rounds an integer to the nearest digit, specified by <code>pos</code>.
+     *
+     * @param num the num the number that's getting rounded.
+     * @param pos the pos the digit to round to.
+     * @return an int which is the rounded number.
+     */
+    // TODO: Please fix this method this is so cringe lol
     public static int round(int num, int pos) {
         String numString = Integer.toString(num);
         if (pos >= numString.length() - 2)
@@ -36,6 +49,12 @@ public class Maths {
     }
 
 
+    /**
+     * Returns the number of significant figures the String decimal number contains
+     *
+     * @param num the number, in String form
+     * @return an int indicating the number of significant figures the number (as a String) contains
+     */
     public static int numSigFigs(String num) {
         num = num.trim();
         if (!num.matches("[0-9.]+"))
@@ -50,6 +69,13 @@ public class Maths {
         }
     }
 
+    /**
+     * Number in pos int.
+     *
+     * @param number the number
+     * @param index  the index
+     * @return the int
+     */
     public static int numberInPos(int number, int index) {
         return Integer.parseInt(Integer.toString(number).substring(index, index + 1));
     }
@@ -58,54 +84,53 @@ public class Maths {
 
     /**
      * Returns the absolute value of a number.
+     *
      * @param num the number given.
      * @return <code>byte</code>-- The absolute value of byte num.
      */
-    public static byte abs(byte num)
-    {
+    public static byte abs(byte num) {
         return num < 0 ? (byte) -num : num;
     }
 
     /**
      * Returns the absolute value of a number.
+     *
      * @param num the number given.
      * @return <code>short</code>-- The absolute value of short num.
      */
-    public static short abs(short num)
-    {
+    public static short abs(short num) {
         return num < 0 ? (short) -num : num;
     }
 
     /**
      * Returns the absolute value of a number.
+     *
      * @param num the number given.
      * @return <code>int</code>-- The absolute value of int num.
      */
-    public static int abs(int num)
-    {
+    public static int abs(int num) {
         return num < 0 ? -num : num;
     }
 
     /**
      * Returns the absolute value of a number.
+     *
      * @param num the number given.
      * @return <code>long</code>-- The absolute value of long num.
      */
-    public static long abs(long num)
-    {
-        return num < 0 ? (long) -num : num;
+    public static long abs(long num) {
+        return num < 0 ? -num : num;
     }
 
     /**
      * This method finds the greatest common denominator (gcd) of the two integers provided. The method used to find the
      * gcd of the two numbers is Euclid's Algorithm but with repeated subtraction.
+     *
      * @param a the first integer given
      * @param b the second integer given
-     * @return <code>int</code>-- The greatest common denominator between two integers;
-     * or -1 if the sum of the parameters is greater than Integer.MAX_VALUE.
+     * @return <code>int</code>-- The greatest common denominator between two integers; or -1 if the sum of the parameters is greater than Integer.MAX_VALUE.
      */
-    public static int gcd(int a, int b)
-    {
+    public static int gcd(int a, int b) {
         a = abs(a);
         b = abs(b);
 
@@ -119,8 +144,7 @@ public class Maths {
         if (b == 0)
             return a;
 
-        while (a != b)
-        {
+        while (a != b) {
             if (a > b)
                 a -= b;
             else
